+++ RELEASE_NOTES.md
@@ -23,7 +23,7 @@
 - Manual ordering of tournaments (3.3.1)
 - Prevent manual reordering of 1 tournament (3.3.4)
 - Fixed the tie-break search (3.3.6)
-- Add an option for specifiy the base date for the age categories (3.4.0)
+- Add an option for specify the base date for the age categories (3.4.0)
 
 ## Players
 
@@ -38,6 +38,7 @@
 - Added place cards (3.3.0)
 - Fixed printing of the player table with multiple tournaments (3.3.2)
 - Fixed clipping of tables when printed (3.3.2)
+- Added prize receipts (3.4.0)
 
 ## Screens